Update with latest Ruby, JRuby and ActiveSupport versions
This includes an activesupport5 gemfile for testing against that version. We also exclude that from older versions of Ruby since ActiveSupport 5 doesn't support them. The commented out line for rbx did not exclude the build itself so that actually causes Travis to run another set of builds against Ruby 1.9.3 which we already specify! No need for this.